+++ src/Components/Icon.js
@@ -1,13 +1,5 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-
-// Font Awesome, load all
-import { library } from '@fortawesome/fontawesome-svg-core'
-import { fas } from '@fortawesome/free-solid-svg-icons';
-import { far } from '@fortawesome/free-regular-svg-icons';
-import { fab } from '@fortawesome/free-brands-svg-icons';
-library.add(fas, far, fab);
 
 const Icon = props => {
     const {
@@ -15,24 +7,11 @@ const Icon = props => {
         ...attributes
     } = props;
 
-    const isJSON = string => {
-        try {
-            JSON.parse(string);
-        } catch {
-            return false;
-        }
-        return true;
-    }
-
     const addClasses = string => {
         return string.replace(/<svg/g, `<svg class="${attributes.className}"`)
     }
 
-    return (
-        (isJSON(props.icon))
-            ? <FontAwesomeIcon icon={JSON.parse(icon)} className="m-auto fa-fw" {...attributes} />
-            : <div dangerouslySetInnerHTML={{ __html: addClasses(icon) }} {...attributes} />
-    )
+    return <div dangerouslySetInnerHTML={{ __html: addClasses(icon) }} {...attributes} />
 }
 
 Icon.propTypes = {